refactor: move store into separate file
