+++ ckanext/nhm/theme/assets/scripts/apps/beetle-iiif/src/app.js
@@ -1,31 +1,7 @@
 import Vue from 'vue';
 import App from './App.vue';
-import Vuex from 'vuex';
+import store from './store';
 import { API } from './api';
-
-const rootElement = document.getElementById('beetle-iiif-app');
-const resourceId = rootElement.getAttribute('data-resource-id');
-
-Vue.use(Vuex);
-
-export const SET_RECORD = 'SET_RECORD';
-export const SET_QUERY = 'SET_QUERY';
-
-const store = new Vuex.Store({
-  state: {
-    resourceId: resourceId,
-    query: {},
-    record: null,
-  },
-  mutations: {
-    [SET_RECORD](state, newRecord) {
-      state.record = newRecord;
-    },
-    [SET_QUERY](state, newQuery) {
-      state.query = newQuery;
-    },
-  },
-});
 
 export const api = new API(store.state);
